test(LadderGameTest): 사다리 생성 테스트 LadderTest로 이관, 참가자와 실행결과 개수 검증 테스트 코드 추가
